# openapi.yaml
openapi: 3.0.3
info:
  title: Ticket Booking Service API
  description: |
    Handles ticket reservation, cancellation, and passenger management
    for the Sri Lanka Train Booking System.

    ## Integration Points
    - **Calls** Seat Availability Service (`POST /seats/reserve`) to prevent double booking
    - **Calls** Payment Service (`POST /payments/refund`) on ticket cancellation
    - **Validates tokens** via Train Management Service (`GET /auth/verify`)

  version: 1.0.0

servers:
  - url: http://localhost:8003
    description: Local development
  - url: https://<YOUR_ECS_IP>:8000
    description: AWS ECS Production

tags:
  - name: Tickets
  - name: Bookings
  - name: Passengers
  - name: Health

paths:

  /health:
    get:
      summary: Health check
      tags: [Health]
      responses:
        "200":
          description: Service healthy
          content:
            application/json:
              example: { status: healthy, service: ticket-booking }

  /tickets/book:
    post:
      summary: Book a new ticket
      tags: [Tickets]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketBookRequest"
            example:
              train_id: TRN-001
              seat_id: S-12A
              seat_class: "2nd Class"
              departure: "Colombo Fort"
              destination: "Kandy"
              departure_time: "2026-03-20T08:00:00"
              arrival_time: "2026-03-20T11:30:00"
              price: 350.00
              passenger:
                name: "Ashan Perera"
                email: "ashan@example.com"
                nic: "199012345678"
                phone: "+94771234567"
      responses:
        "201":
          description: Ticket booked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
        "401":
          description: Unauthorized
        "409":
          description: Seat no longer available

  /tickets/{ticket_id}:
    get:
      summary: Get ticket by ID
      tags: [Tickets]
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
        "404":
          description: Ticket not found

    delete:
      summary: Cancel a ticket
      tags: [Tickets]
      security:
        - BearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  default: "Cancelled by user"
      responses:
        "200":
          description: Ticket cancelled
        "400":
          description: Cannot cancel (already cancelled or completed)
        "404":
          description: Ticket not found

  /tickets/user/{user_id}:
    get:
      summary: Get all tickets for a user
      tags: [Tickets]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
      responses:
        "200":
          description: List of user tickets

  /bookings/{booking_ref}:
    get:
      summary: Get booking by reference
      tags: [Bookings]
      parameters:
        - name: booking_ref
          in: path
          required: true
          schema:
            type: string
            example: BK-00001
      responses:
        "200":
          description: Booking record
        "404":
          description: Booking not found

  /passengers/{passenger_id}:
    get:
      summary: Get passenger info
      tags: [Passengers]
      security:
        - BearerAuth: []
      parameters:
        - name: passenger_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Passenger record
        "404":
          description: Passenger not found

components:
  schemas:
    TicketBookRequest:
      type: object
      required: [train_id, seat_id, seat_class, departure, destination, departure_time, arrival_time, price, passenger]
      properties:
        train_id:       { type: string }
        seat_id:        { type: string }
        seat_class:     { type: string, enum: ["1st Class", "2nd Class", "Sleeper"] }
        departure:      { type: string }
        destination:    { type: string }
        departure_time: { type: string, format: date-time }
        arrival_time:   { type: string, format: date-time }
        price:          { type: number, minimum: 0 }
        passenger:      { $ref: "#/components/schemas/PassengerCreate" }

    PassengerCreate:
      type: object
      required: [name, email, nic, phone]
      properties:
        name:  { type: string }
        email: { type: string, format: email }
        nic:   { type: string }
        phone: { type: string }

    TicketResponse:
      type: object
      properties:
        id:             { type: string }
        train_id:       { type: string }
        seat_id:        { type: string }
        passenger_id:   { type: string }
        user_id:        { type: string }
        seat_class:     { type: string }
        status:         { type: string, enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED] }
        departure:      { type: string }
        destination:    { type: string }
        departure_time: { type: string }
        arrival_time:   { type: string }
        price:          { type: number }
        booking_ref:    { type: string }
        created_at:     { type: string }

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT